First pass of adding OMNeTpp 6 support
Mostly changed use of longValue to intValue in cPars, as the interface
for cPars has changed. Also rectified an issue where an iterator was not
correctly passed to another owner.

+++ policies/DIF/FA/MultilevelQoS/QoSIdComparer/QoSIdComparer.cc
@@ -25,8 +25,8 @@
 Define_Module(QoSIdComparer);
 
 void QoSIdComparer::initialize() {
-    mHops = par("maxHops").longValue();
-    mlBandw = par("mulBandw").longValue();
+    mHops = par("maxHops").intValue();
+    mlBandw = par("mulBandw").intValue();
     setPolicyDisplayString(this);
 }
